updated files

+++ src/main/java/com/schools/school/service/serviceImpl/FeesPaymentServiceImpl.java
@@ -5,6 +5,7 @@ import com.schools.school.repository.FeesPaymentRepository;
 import com.schools.school.service.FeesPaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 import java.util.Optional;
@@ -59,11 +60,27 @@ public class FeesPaymentServiceImpl implements FeesPaymentService {
 
     @Override
     public void updateFeesPayment(FeesPayment feesPayment) {
-        feesPaymentRepository.save(feesPayment);
+
+    }
+
+    @Override
+    public void updateFeesPayment(@PathVariable long id, FeesPayment feesPayment) {
+        FeesPayment existingPayment = feesPaymentRepository.findById(id).get();
+        existingPayment.setFirstName(feesPayment.getFirstName());
+        existingPayment.setMinName(feesPayment.getMinName());
+        existingPayment.setLastName(feesPayment.getLastName());
+        existingPayment.setClassRoom(feesPayment.getClassRoom());
+        existingPayment.setFeesOfPayment(feesPayment.getFeesOfPayment());
+        existingPayment.setFeesPaid(feesPayment.getFeesPaid());
+        existingPayment.setFeeBalance(feesPayment.getFeeBalance());
+        feesPaymentRepository.save(existingPayment);
     }
 
     @Override
     public void deleteFeesPayment(Long id) {
+        if (!feesPaymentRepository.existsById(id)){
+            throw new IllegalArgumentException("No payment made by the id");
+        }
         feesPaymentRepository.deleteById(id);
     }
 }
